+++ src/components/Sidebar/SidebarHeader/SidebarHeader.tsx
@@ -5,11 +5,22 @@ import { CommonAndHTMLProps } from "../../Element/constants";
 
 import { SidebarHeaderStyled } from "./SidebarHeader.styled";
 
+// prettier-ignore
+export interface SidebarHeaderCustomProps {
+    isSticky ? : boolean;
+}
+
 export type SidebarHeaderElementType = HTMLDivElement;
-export type SidebarHeaderProps = CommonAndHTMLProps<SidebarHeaderElementType>;
+export type SidebarHeaderProps = CommonAndHTMLProps<SidebarHeaderElementType> & SidebarHeaderCustomProps;
 
 export const SidebarHeader = React.forwardRef(
-    ({ ...props }: SidebarHeaderProps, ref: React.Ref<SidebarHeaderElementType>) => {
-        return <Element<SidebarHeaderElementType> as={SidebarHeaderStyled} ref={ref} {...props} />;
+    ({isSticky, ...props }: SidebarHeaderProps, ref: React.Ref<SidebarHeaderElementType>) => {
+        let classNames = [];
+
+        if (isSticky) {
+            classNames.push("is-sticky");
+        }
+
+        return <Element<SidebarHeaderElementType> as={SidebarHeaderStyled} ref={ref} classNames={classNames} {...props} />;
     }
 );