+++ src/components/Sidebar/SidebarItemsGroupHeader/SidebarItemsGroupHeader.styled.tsx
@@ -3,5 +3,5 @@ import styled from "styled-components";
 import { SidebarItemsGroupHeaderProps } from "./SidebarItemsGroupHeader";
 
 export const SidebarItemsGroupHeaderStyled = styled.summary`
-    font-weight : ${(props: SidebarItemsGroupHeaderProps) => props.theme.sidebar.subLinks.header.weight};
+    font-weight : ${(props: SidebarItemsGroupHeaderProps) => props.theme.sidebar.linksWrapper.subLinks.header.weight};
 `;