+++ src/components/Sidebar/SidebarLinksWrapper/SidebarLinksWrapper.styled.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { SidebarLinksWrapperProps } from "./SidebarLinksWrapper";
+
+
+export const SidebarLinksWrapperStyled = styled.footer`
+    margin-top    : ${(props: SidebarLinksWrapperProps) => props.theme.sidebar.linksWrapper.marginTop};
+    margin-bottom : ${(props: SidebarLinksWrapperProps) => props.theme.sidebar.linksWrapper.marginBottom};
+`;